# tools: improve js linter