### Pull Request check-list

_Please make sure to review and check all of these items:_

- [X] Does `make -j8 test` (UNIX) or `vcbuild test nosign` (Windows) pass with
  this change (including linting)?
- [X] Is the commit message formatted according to [CONTRIBUTING.md][0]?
- [X] If this change fixes a bug (or a performance problem), is a regression
  test (or a benchmark) included?
- [X] Is a documentation update included (if this change modifies
  existing APIs, or introduces new ones)?

### Affected core subsystem(s)

* `tools`

[0]: https://github.com/nodejs/node/blob/master/CONTRIBUTING.md#step-3-commit

### Description of change

This commit switches from the `eslint` command-line tool to a custom tool that uses `eslint` programmatically in order to perform linting in parallel and to display linting results incrementally instead of buffering them until the end. There is also now a progress indicator similar to that of `make test`.

Additionally, there is a new `lint-ci` rule that disables the progress indicator. I guess the command used on the linter Jenkins node will need to be changed internally as I couldn't find where it's configured to use what command.

Fixes: https://github.com/nodejs/node/issues/5596