@jbergstroem 2016-03-10
We currently do `gmake lint`; so I'll change that to `gmake lint-ci` at some stage then?