@jbergstroem 2016-03-10
@mscdex btw, it's on my agenda to – at some stage – get tap output from the linter so we can post feedback to github. It acts differently than the test suite since lint only reports errors and we'd have an "error count"; but it's still better than no output at all. This may or may not affect your current work but I thought I'd just bring it up.

You can find my work and playing around here: https://github.com/jbergstroem/node/pull/1