@jbergstroem 2016-03-10
@mscdex the problem with switching what command the linter should run is that we need to merge that commit to all branches and have it sit there for a while before we can switch in CI; otherwise it would call a command that wouldn't exist. My plan to address this was introducing placeholders and let them sit around for a while. I obviously forgot about it.