@jbergstroem 2016-03-10
@mscdex I recall globbing files ("collect `*.tap`") worked just fine.