@mscdex 2016-03-10
Ok, I've updated it to allow output to file and selecting the formatter.