@silverwind 2016-03-10
Nice work, seems to run slightly faster too:

Before: `make jslint  15.72s user 0.82s system 106% cpu 15.565 total`
After: `make jslint  33.02s user 2.56s system 287% cpu 12.388 total`