@silverwind 2016-03-10
3s faster, but a lot more cpu intensive.