@mscdex 2016-03-10
I am getting ~7s improvement on my machine with the current list of files we lint.

I also just bumped up the max load per worker and that seemed to decrease times a tad more.

It will use more cpu time if you have multiple cores though.