@jbergstroem 2016-03-10
I again just want to make the note that we can't land this as-is seeing how we rely on `make lint` from ci (just want this here in case of LGTM's starting to roll in). Few options:

1. New PR with `make lint-ci` calling `make lint`, merge it and wait a few months before landing this
2. Env logic (since its easy for us to add env stuff through jenkins) for `make lint` -- this is how we usually get around handling multiple versions of `Makefile`.
3. I had something in mind but forgot while writing the other two :) will revisit when  awake.