@bnoordhuis 2016-03-23
LGTM although there are a few conflicts.  @jbergstroem What needs to happen before this can land?