@jbergstroem 2016-03-23
@bnoordhuis I suggest we do the first path; create placeholders in Makefile and make sure we get that commit into all branches we want to cover in CI, then land this.