@Trott 2016-03-23
I don't think the `make lint-ci` job is not working correctly. If I modify `.eslintrc` in the root directory (say, remove the option from`comma-dangle` so that it is just `2`), `make jslint` correctly flags errors but `make lint-ci` reports 0 errors.