@Trott 2016-03-23
Nit: I'm not too stoked on the way the error output is interspersed with the progress indicator:

```
[00:09|% 58|+   505|-    55]: test
/Users/trott/io.js/test/parallel/test-zlib-convenience-methods.js
  13:18  error  Unexpected trailing comma  comma-dangle
  20:31  error  Unexpected trailing comma  comma-dangle

✖ 2 problems (2 errors, 0 warnings)

[00:09|% 58|+   526|-    56]: test
/Users/trott/io.js/test/sequential/test-net-localport.js
  17:31  error  Unexpected trailing comma  comma-dangle

✖ 1 problem (1 error, 0 warnings)

[00:09|%100|+   584|-    57]: Done
make: *** [jslint] Error 1
```

This will break my workflow for new rules as I often do `make jslint > some-file` so I can process the file later, but the end result will be like this:

```


[00:00|%  0|+     0|-     0]: Starting ...
                                          
[00:01|%100|+     0|-     0]: lib
                                 
[00:02|%100|+     7|-     0]: lib
/Users/trott/io.js/lib/internal/child_process.js
  572:31  error  Unexpected trailing comma  comma-dangle
  601:25  error  Unexpected trailing comma  comma-dangle

✖ 2 problems (2 errors, 0 warnings)


                                 
[00:02|%100|+     7|-     1]: lib
/Users/trott/io.js/lib/cluster.js
  650:19  error  Unexpected trailing comma  comma-dangle

✖ 1 problem (1 error, 0 warnings)


                                 
[00:02|%100|+    15|-     2]: lib
                                 
[00:03|%100|+    40|-     2]: lib
                                 
[00:03|%100|+    44|-     2]: Done
```

I'm labeling this a nit because there are obvious workarounds: Run `eslint` directly, use the CI linting rule instead, probably other solutions. But if it's not a big deal to have the progress indicator go to `STDERR` while the error reporting goes to `STDOUT`, that would help me personally. Not sure if it causes other problems or violates a sense of where the right places for things are.