@Trott 2016-03-23
I'm deleting my previous comment about the formatting of the make job when it's redirected. It's really the least of my worries. The speed increase is a good trade-off.