@mscdex 2016-03-26
@Trott That message is from `cpplint`, not `jslint-ci`.

@jbergstroem What would that kind of initial change look like? I'd like to get the placeholders in sooner than later.