@jbergstroem 2016-03-26
@mscdex I'll do a PR.