@Trott 2016-03-26
@mscdex Ah, yes, you are correct, of course. The errors are showing up just fine in the `tap` file.