@Fishrock123 2016-03-28
@mscdex could you comment-document the new jslint proxy tool more? I just want to make sure this isn;t something we have to fight with much in the future. :)