@silverwind 2016-04-09
Has there been thoughts about getting something like this into ESLint itself? Maybe through `--parallel` or similar.