@mscdex 2016-04-09
@silverwind Not just parallelization but also (nice) progress output.