@jbergstroem 2016-04-15
@mscdex now that we've landed the new targets in our 4.x branch as well we could perhaps look at updating this? After landing we can update jenkins to use the new targets.