@jbergstroem 2016-04-15
By "update" I just meant look at merging; sorry. I've changed targets in ci now -- this might imply that old jobs that hasn't been rebased as of late will likely fail; so lets try and keep track of that. Failed lint jobs aren't common, so it should be pretty easy to spot. LGTM.