@jbergstroem 2016-04-15
Actually; hold off passing `-J` in test-ci for now. I'd like to add it similar to how its done in https://github.com/nodejs/node/pull/6208.