@jbergstroem 2016-04-15
FYI: Went from spending >50sec to <30sec in our CI on linting as a result of this PR 🎉