@TheAlphaNerd 2016-04-21
@mscdex do you think we should backport this to v4?