@mscdex 2016-04-21
@TheAlphaNerd AFAIK it should work as long as a2ca347803 is also applied.