@TheAlphaNerd 2016-05-17
I'm going to mark this as do not land for lts right now

@mscdex feel free to prepare a backport of this change if you think it would be beneficial